chore: update @bitcoinbrisbane/block52 to version 1.0.57 and refactor Deck model to use imported Card and SUIT types
