+++ sdk/src/types/game.ts
@@ -84,3 +84,17 @@ export type TexasHoldemStateDTO = {
     winners: WinnerDTO[];
     signature: string;
 };
+
+export enum SUIT {
+    CLUBS = 1,
+    DIAMONDS = 2,
+    HEARTS = 3,
+    SPADES = 4
+}
+
+export type Card = {
+    suit: SUIT;
+    rank: number;
+    value: number;
+    mnemonic: string;
+};
